refactor: relocate EsntlIdGenerator to util package
